implement getWithDefault
