+++ types/ember/test/observable.ts
@@ -1,4 +1,5 @@
 import Ember from 'ember';
+import { assertType } from './lib/assert';
 
 class MyComponent extends Ember.Component {
     foo = 'bar';
@@ -18,4 +19,66 @@ class MyComponent extends Ember.Component {
     fooDidChange(sender: MyComponent, key: 'foo') {
         // your code
     }
+}
+
+const person = Ember.Object.create({
+    name: 'Fred',
+    age: 29,
+    capitalized: Ember.computed<string>(function() {
+        return this.get('name').toUpperCase();
+    })
+});
+
+const pojo = { name: 'Fred', age: 29 };
+
+function testGet() {
+    assertType<string>(Ember.get(person, 'name'));
+    assertType<number>(Ember.get(person, 'age'));
+    assertType<string>(Ember.get(person, 'capitalized'));
+    assertType<string>(person.get('name'));
+    assertType<number>(person.get('age'));
+    assertType<string>(person.get('capitalized'));
+    assertType<string>(Ember.get(pojo, 'name'));
+}
+
+function testGetProperties() {
+    assertType<{ name: string }>(Ember.getProperties(person, 'name'));
+    assertType<{ name: string, age: number }>(Ember.getProperties(person, 'name', 'age'));
+    assertType<{ name: string, age: number }>(Ember.getProperties(person, [ 'name', 'age' ]));
+    assertType<{ name: string, age: number, capitalized: string }>(Ember.getProperties(person, 'name', 'age', 'capitalized'));
+    assertType<{ name: string }>(person.getProperties('name'));
+    assertType<{ name: string, age: number }>(person.getProperties('name', 'age'));
+    assertType<{ name: string, age: number }>(person.getProperties([ 'name', 'age' ]));
+    assertType<{ name: string, age: number, capitalized: string }>(person.getProperties('name', 'age', 'capitalized'));
+    assertType<{ name: string, age: number }>(Ember.getProperties(pojo, 'name', 'age'));
+}
+
+function testGetWithDefault() {
+    assertType<string>(Ember.getWithDefault(person, 'name', 'Joe'));
+    assertType<number>(Ember.getWithDefault(person, 'age', 20));
+    assertType<string>(Ember.getWithDefault(person, 'capitalized', 'JOE'));
+    assertType<string>(person.getWithDefault('name', 'Joe'));
+    assertType<number>(person.getWithDefault('age', 20));
+    assertType<string>(person.getWithDefault('capitalized', 'JOE'));
+    assertType<string>(Ember.getWithDefault(pojo, 'name', 'JOE'));
+}
+
+function testSet() {
+    assertType<string>(Ember.set(person, 'name', 'Joe'));
+    assertType<number>(Ember.set(person, 'age', 35));
+    assertType<string>(Ember.set(person, 'capitalized', 'JOE'));
+    assertType<string>(person.set('name', 'Joe'));
+    assertType<number>(person.set('age', 35));
+    assertType<string>(person.set('capitalized', 'JOE'));
+    assertType<string>(Ember.set(pojo, 'name', 'Joe'));
+}
+
+function testSetProperties() {
+    assertType<{ name: string }>(Ember.setProperties(person, { name: 'Joe' }));
+    assertType<{ name: string, age: number }>(Ember.setProperties(person, { name: 'Joe', age: 35 }));
+    assertType<{ name: string, capitalized: string }>(Ember.setProperties(person, { name: 'Joe', capitalized: 'JOE' }));
+    assertType<{ name: string }>(person.setProperties({ name: 'Joe' }));
+    assertType<{ name: string, age: number }>(person.setProperties({ name: 'Joe', age: 35 }));
+    assertType<{ name: string, capitalized: string }>(person.setProperties({ name: 'Joe', capitalized: 'JOE' }));
+    assertType<{ name: string, age: number }>(Ember.setProperties(pojo, { name: 'Joe', age: 35 }));
 }
