+++ types/ember/tsconfig.json
@@ -29,11 +29,7 @@
         "test/array.ts",
         "test/array-ext.ts",
         "test/array-proxy.ts",
-        "test/get.ts",
-        "test/get-properties.ts",
         "test/helper.ts",
-        "test/set.ts",
-        "test/set-properties.ts",
         "test/computed.ts",
         "test/component.ts",
         "test/function-ext.ts",
